% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{X}}{matrix.}

\item{\code{Y}}{matrix.}

\item{\code{reg_coe}}{vector.}

\item{\code{fit_val}}{vector.}

\item{\code{resid_e}}{vector.}

\item{\code{n}}{numeric.}

\item{\code{p}}{numeric.}

\item{\code{df}}{numeric.}

\item{\code{resid_var}}{numeric.}

\item{\code{var_beta}}{matrix.}

\item{\code{t_val}}{vector.}
}}


